+++ start_crisp.py
@@ -12,10 +12,101 @@
 ###
 """
 
+
 import sys, os, time, random, subprocess, balance_queue, shutil
 from os import path as op
 from coadaptree import makedir, fs, pklload, get_email_info
-from combine_crisp import getfiles, check_queue, check_seff
+from balance_queue import getsq
+
+
+def gettimestamp(f):
+    return time.ctime(op.getmtime(f))
+
+
+def getmostrecent(files, remove=False):
+    if not isinstance(files, list):
+        files = [files]
+    if len(files) > 1:
+        whichout = files[0]
+        dt1 = dt.strptime(gettimestamp(whichout)[4:], "%b %d %H:%M:%S %Y")
+        for o in files[1:]:
+            dt2 = dt.strptime(gettimestamp(o)[4:], "%b %d %H:%M:%S %Y")
+            if dt2 > dt1:
+                whichout = o
+                dt1 = dt2
+        if remove is True:
+            [os.remove(f) for f in files if not f == whichout]
+        return whichout
+    elif len(files) == 1:
+        return files[0]
+    else:
+        # if len(files) == 0
+        return None
+
+
+def getfiles(samps, shdir, grep):
+    if grep == 'crisp_bedfile':
+        # numbedfiles = numshfiles = found; so set samps to numbedfiles to bypass len(found) == len(samps)
+        pooldir = op.dirname(op.dirname(shdir))
+        parentdir = op.dirname(pooldir)
+        pool = op.basename(pooldir)
+        ref = pklload(op.join(parentdir, 'poolref.pkl'))[pool]
+        samps = fs(op.join(op.dirname(ref), 'bedfiles_%s' % op.basename(ref).replace(".fasta","")))
+    found = [sh for sh in fs(shdir) if sh.endswith(".sh") and grep in sh]
+    outs = [out for out in fs(shdir) if out.endswith('.out') and grep in out]
+    if len(found) != len(samps):
+        # this only works for start crisp, not combine crisp since crisp has 15 bedfiles.
+        print('not all shfiles have been created, exiting %s' % sys.argv[0])
+        exit()
+    files = dict((f, getmostrecent([out for out in outs if op.basename(f).replace(".sh", "") in out]))
+                 for f in found)
+    if None in files.values():
+        print('not all shfiles have been sbatched, exiting %s' % sys.argv[0])
+        exit()
+    return files
+
+
+def check_seff(outs):
+    jobid = os.environ['SLURM_JOB_ID']
+    for f in outs:
+        pid = f.split("_")[-1].replace(".out", "")
+        if not pid == jobid:
+            seff, seffcount = '', 0
+            while isinstance(seff, list) is False:
+                # sometimes slurm sucks
+                seff = subprocess.check_output(['seff', pid]).decode('utf-8').split('\n')
+                if seffcount == 10:
+                    print('slurm is screwing something up with seff, exiting %s' % sys.argv[0])
+                    exit()
+                time.sleep(1)
+                seffcount += 1
+            state = [x.lower() for x in seff if 'State' in x][0]
+            if 'exit code 0' not in state:
+                text = 'died' if not 'running' in state else 'running'
+                print('cannot proceed with %s' % sys.argv[0])
+                print('job died (%s) for %s' % (state, f)) 
+                print('exiting %s' % sys.argv[0])
+                exit()
+
+
+def checkpids(outs, queue):
+    # if any of the other crisp jobs are pending or running, exit
+    pids = [q[0] for q in queue]
+    jobid = os.environ['SLURM_JOB_ID']
+    for out in outs:
+        pid = out.split("_")[-1].replace(".out", "")
+        if pid in pids and pid != jobid:  # if the job is running, but it's not this job
+            print('the following file is still in the queue - exiting %s' % sys.argv[0],
+                  '\n', '\t%(out)s' % locals())
+            exit()  # TODO: should I be exiting?
+
+
+def check_queue(outs, pooldir):
+    # get jobs from the queue, except those that are closing (assumes jobs haven't failed)
+    sq = getsq(grepping=['crisp_bedfile', op.basename(pooldir)], states=['R', 'PD'])
+    if len(sq) > 0:
+        checkpids(outs, sq)
+    # no need for an else statement here, if len(sq) == 0: no need to check the pids
 
 
 def get_bamfiles(samps, pooldir):
@@ -116,9 +207,6 @@ gatk VariantsToTable --variant {outfile} -F CHROM -F POS -F REF -F ALT -F AF -F 
 -O {tablefile} --split-multi-allelic
 module unload gatk
 
-# combine crisp files
-python $HOME/pipeline/combine_crisp.py {pooldir}
-
 # gzip outfiles to save space
 cd $(dirname {outfile})
 gzip {outfile}
@@ -133,9 +221,9 @@ rm {logfile}
 
 def sbatch(file):
     os.chdir(op.dirname(file))
-    subprocess.call([shutil.which('sbatch'), file])
-    # os.system('sbatch %s' % file)
+    pid = subprocess.check_output([shutil.which('sbatch'), file]).decode('utf-8').replace("\n", "").split()[-1]
     print("sbatched %s" % file)
+    return pid
 
 
 def get_bedfiles():
@@ -146,10 +234,44 @@ def get_bedfiles():
 
 def create_sh(bamfiles, crispdir, pool, pooldir):
     bedfiles = get_bedfiles()
+    pids = []
     for bedfile in bedfiles:
         file = make_sh(bamfiles, bedfile, crispdir, pool, pooldir)
-        sbatch(file)
-    print("done sbatching, exiting %s" % sys.argv[0])
+        pids.append(sbatch(file))
+    return pids
+
+
+def create_combine(pids, parentdir, pool):
+    pooldir = op.join(parentdir, pool)
+    email_text = get_email_info(parentdir,'crisp')
+    if isinstance(pids, list) is not True:
+        print('pids is not a list')
+        pids = [pids]
+    dependencies = '#SBATCH --dependency=afterok:' + ','.join(pids)
+    text = f'''#!/bin/bash
+#SBATCH --job-name={pool}-combine-crisp
+#SBATCH --time=02:59:00
+#SBATCH --mem=8000M
+#SBATCH --cpus-per-task=1
+#SBATCH --output={pool}-combine-crisp_%j.out
+{dependencies}
+{email_text}
+
+
+source $HOME/.bashrc
+export PYTHONPATH="${{PYTHONPATH}}:$HOME/pipeline"
+export SQUEUE_FORMAT="%.8i %.8u %.12a %.68j %.3t %16S %.10L %.5D %.4C %.6b %.7m %N (%r)"
+
+python $HOME/pipeline/combine_crisp.py {pooldir}
+
+'''
+    pooldir = op.join(parentdir, pool)
+    shdir = op.join(pooldir, 'shfiles/crisp')
+    combfile = op.join(shdir, '%s-combine-crisp.sh' % pool)
+    with open(combfile, 'w') as o:
+        o.write("%s" % text)
+    sbatch(combfile)
+    print('sbatched combinefile with dependencies: ' + ','.join(pids))
 
 
 def main(parentdir, pool):
@@ -162,14 +284,13 @@ def main(parentdir, pool):
     crispdir = create_reservation(op.join(parentdir, pool))
 
     # create .sh file and submit to scheduler
-    create_sh(bamfiles, crispdir, pool, op.join(parentdir, pool))
+    pids = create_sh(bamfiles, crispdir, pool, op.join(parentdir, pool))
+    
+    # create .sh file to combine crisp parallels using jobIDs as dependencies
+    create_combine(pids, parentdir, pool)
 
     # balance queue
     balance_queue.main('balance_queue.py', 'crisp')
-    # os.system('python $HOME/pipeline/balance_queue.py crisp')
-
-    # cancel reservation
-    # I don't think I need to do this unless I find that jobs are dying unexpectedly
 
 
 if __name__ == "__main__":
